requirements.txt
streamlit
pymupdf
google-generativeai
scikit-learn